refactor: move command handlers to its own files, remove command handlers classes
