+++ src/commands/compression/compress.js
@@ -1,0 +1,7 @@
+import { createBrotliCompress } from "zlib";
+import { handleCompression } from "./common.js";
+
+export const compress = (args) => {
+  const brotli = createBrotliCompress();
+  handleCompression(brotli, args);
+};
