+++ src/commands/compression/index.js
@@ -1,0 +1,2 @@
+export * from "./compress.js";
+export * from "./decompress.js";
