+++ src/commands/crypto/index.js
@@ -1,0 +1,1 @@
+export * from "./hash.js";
