+++ src/commands/fileSystem/index.js
@@ -1,0 +1,9 @@
+export * from "./add.js";
+export * from "./changeDirectory.js";
+export * from "./copy.js";
+export * from "./list.js";
+export * from "./move.js";
+export * from "./print.js";
+export * from "./remove.js";
+export * from "./rename.js";
+export * from "./up.js";
