+++ src/commands/fileSystem/remove.js
@@ -1,0 +1,6 @@
+import { rm } from "fs/promises";
+
+export const remove = async (args) => {
+  const [pathToFile] = args;
+  await rm(pathToFile);
+};
