+++ src/commands/fileSystem/up.js
@@ -1,0 +1,11 @@
+import { join, parse } from "path";
+import { pathStorage } from "../../storages/PathStorage.js";
+
+const isRootPath = (currentPath) => {
+  return currentPath === parse(currentPath).root;
+};
+
+export const up = () => {
+  const path = isRootPath(pathStorage.get()) ? pathStorage.get() : join(pathStorage.get(), "..");
+  pathStorage.set(path);
+};
