+++ src/commands/process/exit.js
@@ -1,0 +1,3 @@
+export const exit = () => {
+  process.emit("SIGINT");
+};
