--- src/handlers/ExitCommandHandler.js
@@ -1,5 +1,0 @@
-export class ExitCommandHandler {
-  handle() {
-    process.emit("SIGINT");
-  }
-}
